Removed deprecated *Session* artifacts.
