+++ Source/Otc.AuthorizationContext.Abstractions/Otc.AuthorizationContext.Abstractions.csproj
@@ -4,8 +4,8 @@
     <TargetFramework>netstandard2.0</TargetFramework>
     <Authors>Ole Consignado</Authors>
     <Copyright>Ole Consignado (c) 2018</Copyright>
-    <VersionPrefix>1.0.1</VersionPrefix>
-    <PackageProjectUrl>https://github.com/OleConsignado/otc-session-context</PackageProjectUrl>
+    <VersionPrefix>1.0.0</VersionPrefix>
+    <PackageProjectUrl>https://github.com/OleConsignado/otc-authorization-context</PackageProjectUrl>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
